feat: Add async function to fetch feeds and pass data to FeedContaine… (#7)
* feat: Add async function to fetch feeds and pass data to FeedContainer component

- Added async function to fetch feeds in feedsPage component
- Passed fetched feeds data to FeedContainer component
- Updated FeedContainer component to receive feedData as props
- Updated SERVER_URL to use process.env.SERVER_URL or fallback to 'http://localhost:8080' in FeedsApi.ts and CommentApi.ts

* feat: Add Mock Data

* fix: Update getSessionToken to be an async function in CommentApi and FeedsApi

Updated the getSessionToken function in CommentApi and FeedsApi to be an async function to ensure that the session token is retrieved correctly before making API requests. This change ensures that the session token is obtained in a synchronous manner, preventing potential race conditions and ensuring the security of the API calls.